requirements.txt
asgiref==3.3.4
astpretty==2.1.0
attrs==21.2.0
cognitive-complexity==1.2.0
Django==3.2.3
django-crispy-forms==1.11.2
eradicate==2.0.0
flake8==3.9.2
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.5.0
flake8-django==1.1.2
flake8-eradicate==1.0.0
flake8-expression-complexity==0.0.9
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-variables-names==0.0.4
gunicorn==20.0.4
mccabe==0.6.1
pep8-naming==0.11.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
